requirements.txt
arc4==0.4.0
pycryptodome==3.17
urllib3==1.26.15
uvicorn==0.21.1
